requirements.txt
# requirements.txt
# FastAPI y Servidores
fastapi
uvicorn
gunicorn

# Base de Datos y ORM
sqlalchemy==2.0.31
psycopg2-binary
alembic

# Configuración
pydantic-settings
python-dotenv

# Seguridad y Autenticación
passlib[bcrypt]
python-jose[cryptography]
authlib

# Analytics y HTTP Clients
httpx
pydantic[email]

# Session Management
starlette